amend varnish vcl_fetch not to process the response on non GET
If the request is made using post, the headers should be passed to
the client. One common use case is the login functionality, where
Magento will change the session cookie.

Another option would be to make POST requests to be piped instead of
passed. While it would not require checking the fetched response,
it would change the statistics and would also bypass the varnish
timeout (error) handling mechanism.