Fix type conflit between defaults and validation
The URI builder validates the port parameter to make sure it is an integer.
Meanwhile, the default values are strings. This causes the recipe to fail,
and breaks compatibility with old configurations where the values are string
as well.
This commit introduces a conversion to int on the port, to make sure it passed
validation.