Stop solr from deleting/re-creating solr config.
"/etc/solr/conf/" was being deleted upon each provision and new config
copied from the "files/solr.config/" directory.

This makes it hard to see actual changes to configuration on opscode
and slows down the dev VM provisioning by 6-10 seconds.

This commit changes the approach to be a wrapper cookbook instead of placing
the files itself.